requirements.txt

crewai
crewai-tools
langchain
langchain_community
redis
chromadb
langchain-chroma
langchain-openai
numpy
pandas
ipython
pypdf
# FastAPI and web framework dependencies
fastapi
requests
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
reportlab
# Database dependencies
sqlalchemy
alembic
psycopg2-binary
pymongo>=4.0.0

boto3
Pillow

emails

pydantic
pydantic-settings
email-validator
markdown>=3.4.0

matplotlib>=3.7.0
seaborn>=0.12.0

# YAML support
PyYAML

# Async MongoDB
motor>=3.3.1

# PDF text extraction (assignment briefs)
pypdf>=4.0.0
python-dateutil>=2.8.0

# WhatsApp Cloud API HTTP client
httpx>=0.27.0